requirements.txt
numpy<2.0
pandas
matplotlib
cmdstanpy
scipy
seaborn
torch
plotly
scikit-learn
pyro-ppl
tqdm
